Creation of user metadata, login & register logic update and customer profile page
System can now recognize whether it's a customer or a merchant logging in depending on the toggle selected in the login. The system will look for the respective email within the customer or merchant collection depending on the toggle. This means that one email account can be both a customer or a merchant. Also handled the routing logic for both customer and merchant such that if profile details are not complete, the user or merchant will always be redirected to that page upon login to fill in the rest of the details. Also created the CustomerProfile page with some minor tweaks from the intended design.