Checkout Functionality
Added a balance attribute for merchants and customers to track their current money balance.
If balance is sufficient to pay for total price of cart,
Successful checkout, then
i) update user's balance in Firebase
ii) update merchant's quantity of listing and merchant's balance in Firebase
iii) create an order containing all the listings/products in the cart
iv) empty the cart
v) redirect to order summary page

Order summary page is to display the newly created order.

If the balance is insufficient to pay for cart's items, throw an alert, and redirect to the wallet page.